requirements.txt
langchain
langchain-groq
langchain-core
langchain-community
streamlit
scikit-learn
xgboost
pandas
numpy
spacy
PyPDF2
python-docx
beautifulsoup4
python-dotenv
joblib